refactor: 컴포넌트 분리(지도, 정렬, 매물리스트)
